Fix for bug in .num_expected_args() when `fun` is supplied that has any default argument values of `NA` or `NULL`
Addresses https://github.com/isciences/exactextractr/issues/110

exactextractr::exact_extract() errors out when a `fun` is supplied that has any default argument values of `NA` or `NULL`.

We can trace this back to this internal helper function, `.num_expected_args()`, which counts the number of arguments that {exactextractr} expects a user to supply (i.e., not counting arguments that have a default value, as far as I can tell). But the logical operation of `nchar(a) == 0` will produce `NA` instead of a count if any arguments have a default value of `NA` or `NULL`. Including `na.rm = TRUE` appropriately ignores arguments that have a default value (i.e., counts them as `0` in the total) but produces an integer count rather than an `NA`. Including `na.rm = TRUE` also appropriately results in a count of `0` if _all_ arguments have defaults of `NA` or `NULL`.